+++ src/app/services/external/google/README.md
@@ -1,0 +1,4 @@
+___________________________________
+Info
+* Google API
+* key:  AIzaSyCqZ3dPYustmQUajzIhYu7MJeJ_ePHPnyk    /   DiveInnFontTest
